tentando resolver as parada, botando fogo no parquinho
