+++ labe-commerce/public/manifest.json
