+++ labe-commerce/src/App.js
@@ -1,0 +1,25 @@
+import React from "react";
+
+//THIS WORKS
+function App() {
+  return (
+    <div className="App">
+      <header />
+      <body>
+        <filtros />
+        <produtos />
+
+
+        <carrinho />
+      </body>
+      <prefooter />
+      <footer />
+    </div>
+  );
+}
+
+
+
+export default App;
+
+
